lemon: applied upstream patch 796930d5fc2036c7, dated 2013-10-11
Original comment: Fix a harmless compiler warning in lemon.c.

+++ src/lemon/lemon.c
@@ -3491,7 +3491,7 @@ void print_stack_union(
         break;
       }
       hash++;
-      if( hash>=arraysize ) hash = 0;
+      if( hash>=(unsigned)arraysize ) hash = 0;
     }
     if( types[hash]==0 ){
       sp->dtnum = hash + 1;
